Move pydantic config into separate file.
